add interface.c

+++ src/lib/list.c
@@ -3,19 +3,19 @@
 
 struct llist {
   int size;
-  int * list;
+  long * list;
 };
 
 LONG_list create_list(int size) {
     if(size <= 0) return NULL;
     LONG_list l = malloc(sizeof(struct llist));
     l->size = size;
-    l->list = malloc(sizeof(int) * size);
+    l->list = malloc(sizeof(long) * size);
     return l;
 }
 
 long get_list(LONG_list l, int index) {
-    return l->list[index]; 
+    return l->list[index];
 }
 
 void set_list(LONG_list l, int index, long value) {
@@ -27,4 +27,3 @@ void free_list(LONG_list l) {
         free(l);
     }
 }
-
